requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.1.0
Flask-Babel==3.1.0
WTForms==3.2.1
SQLAlchemy==2.0.43
Werkzeug==3.1.3
qrcode==7.4.2
Pillow==10.2.0
requests==2.31.0
python-dateutil==2.9.0.post0
alembic==1.16.5
python-dotenv==1.0.0
gunicorn==21.2.0